64-bit correctness

+++ appmain/scriptapp.h
@@ -102,7 +102,7 @@ public:
         
     ~EventSink()
     {
-        if (obj && (long)obj != 1)
+        if (obj && (long long)obj != 1)
         {
             obj->unregisterEventListener(this);
         }
